story form update
