--- client/src/components/uploadStoryForm/imageInput.js
@@ -1,79 +1,0 @@
-import React, { useState } from 'react';
-
-function Upload() {
-  const [fileInputState, setFileInputState] = useState('');
-  const [previewSource, setPreviewSource] = useState('');
-  const [selectedFile, setSelectedFile] = useState();
-
-  const previewFile = (file) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onloadend = () => {
-      setPreviewSource(reader.result);
-    };
-  };
-
-  const handleFileInputChange = (e) => {
-    const file = e.target.files[0];
-    previewFile(file);
-    setSelectedFile(file);
-    setFileInputState(e.target.value);
-  };
-
-  const uploadImage = async (base64EncodedImage) => {
-    try {
-      await fetch('/api/v1/story', {
-        method: 'POST',
-        body: JSON.stringify({ data: base64EncodedImage }),
-        headers: { 'Content-Type': 'application/json' },
-      });
-      setFileInputState('');
-      setPreviewSource('');
-    } catch (err) {
-      throw new Error(`Error ${err}`);
-    }
-  };
-  const handleSubmitFile = (e) => {
-    e.preventDefault();
-    if (!selectedFile) return;
-    const reader = new FileReader();
-    reader.readAsDataURL(selectedFile);
-    reader.onloadend = () => {
-      uploadImage(reader.result);
-    };
-    reader.onerror = () => {
-      console.error('AHHHHHHHH!!');
-    };
-  };
-  return (
-    <div>
-      <h1 className="title">Upload an Image</h1>
-      <form onSubmit={handleSubmitFile} className="form">
-        <input
-          id="fileInput"
-          type="file"
-          name="image"
-          onChange={handleFileInputChange}
-          value={fileInputState}
-          className="form-input"
-        />
-        <button className="btn" type="submit">
-          Submit
-        </button>
-      </form>
-      {previewSource && (
-        <img
-          src={previewSource}
-          alt="chosen"
-          style={{ height: '300px' }}
-        />
-      )}
-    </div>
-  );
-}
-
-export default Upload;
-
-/* display flex - (flex column - alin items center)
-border radues - box shadow
-text alin center center - start */
