+++ client/src/components/uploadStoryForm/styles.css
@@ -1,0 +1,13 @@
+.form-input{
+    
+    display: contents;
+}
+.ant-modal-content{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+.form{
+    display: flex;
+    flex-direction: column;
+}
